BT history: restructure components and folders
